Added accumulators to solve for time
Theres tests for it as well. Also, theres a bug fix where the first camera doesn't seem to get recorded in the ptk file.

+++ src/asp/PhotometryTK/tests/TestAlbedoAccumulators.cxx
@@ -58,7 +58,7 @@ protected:
 };
 
 typedef AlbedoAccumulatorTest<uint8> AlbedoAccumulatorTestInt;
-typedef AlbedoAccumulatorTest<uint8> AlbedoAccumulatorTestFloat;
+typedef AlbedoAccumulatorTest<float> AlbedoAccumulatorTestFloat;
 
 #define EXPECT_IMAGE_NULL( p, i ) \
   EXPECT_EQ( p(), i(0,0) );       \
